// packages/alloy-typespec/openapi3-to-typespec/src/openapi-types.ts
export type Extensions = {
  [key in `x-${string}`]?: any;
};

export interface Ref {
  $ref: string;
  description?: string;
  summary?: string;
}

export type Refable<T> = Ref | T;

export type JsonSchemaType =
  | "array"
  | "boolean"
  | "integer"
  | "null"
  | "number"
  | "object"
  | "string";

export type JsonSchema<AdditionalVocabularies extends {} = {}> =
  AdditionalVocabularies & {
    /**
     * The JSON type for the schema.
     * If the value is an array, then an instance validates successfully if its type
     * matches any of the types indicated by the string in the array.
     *
     * @see https://json-schema.org/draft/2020-12/json-schema-validation#name-type
     */
    type?: JsonSchemaType | JsonSchemaType[];
// ...
  };

export type OpenAPISchema3_1 = JsonSchema<
  {
    /**
     * Adds support for polymorphism.
     * The discriminator is an object name that is used to differentiate between other schemas which may satisfy the payload description
     *
     * @see https://spec.openapis.org/oas/v3.1.1.html#discriminator-object
     */
    discriminator?: OpenAPI3Discriminator;

    /**
     * Additional external documentation for this schema.
     *
     * @see https://spec.openapis.org/oas/v3.1.1.html#external-documentation-object
     */
    externalDocs?: OpenAPI3ExternalDocs;

    /**
     * This MAY be used only on properties schemas.
     * It has no effect on root schemas.
     * Adds additional metadata to describe the XML representation of this property.
     */
    xml?: OpenAPI3XmlSchema;
  } & Extensions
>;

export interface OpenAPI3Discriminator extends Extensions {
  propertyName: string;
  mapping?: Record<string, string>;
}

export interface OpenAPI3ExternalDocs {
  url: string;
  description?: string;
}

export type OpenAPI3XmlSchema = Extensions & {
  name?: string;
  namespace?: string;
  prefix?: string;
  attribute?: boolean;
  wrapped?: boolean;
};